Fix non chainable multi line relations
This fixes an issue where relations in dev and test environments were
not chainable across multiple lines. In dev and test environments, after
every line of code executed, `.to_a` is called on relations so that you
can see their output. This was causing the `.loaded?` method to return
true, and any extra queries not being run. This fixes that.