Update server and tests
This updates the mock server to respond properly to the new query
endpoint and query style. It also fixes all tests. Another update is
setting the test environment in accordance with the new style that
ruby-montage accepts. Also updates to the query methods to properly
parse and run the queries.